+++ src/proj6BittingCerratoCohenEllmer/view/CodeTabs.fxml
@@ -12,7 +12,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 
 <TabPane xmlns:fx="http://javafx.com/fxml/1"
-         fx:controller="proj6BittingCerratoCohenEllmer.CodeTabController"
+         fx:controller="proj6BittingCerratoCohenEllmer.controllers.CodeTabController"
          fx:id="tabPane" layoutX="14.0" layoutY="112.0"
          prefHeight="396.0" prefWidth="800.0" tabClosingPolicy="ALL_TABS"
          AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
